+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Verb Changelog
 ## master
 - Removed `verb-max-redirections` variable, as it wasn't working as intended. Use `url-max-redirections` instead.
+- Add `default` argument to `verb-var`.
 
 ## **2.5.0** - 2020-01-30
 - Added `verb-trim-body-end` variable.